+++ rotkehlchen/utils/misc.py
@@ -90,7 +90,7 @@ def timestamp_to_iso8601(ts: Timestamp, utc_as_z: bool = False) -> str:
     return res if utc_as_z is False else res.replace('+00:00', 'Z')
 
 
-def satoshis_to_btc(satoshis: FVal) -> FVal:
+def satoshis_to_btc(satoshis: Union[int, FVal]) -> FVal:
     return satoshis * FVal('0.00000001')
 
 
@@ -228,8 +228,8 @@ def request_get(
 
     try:
         result = rlk_jsonloads(response.text)
-    except json.decoder.JSONDecodeError:
-        raise UnableToDecryptRemoteData(f'{url} returned malformed json')
+    except json.decoder.JSONDecodeError as e:
+        raise UnableToDecryptRemoteData(f'{url} returned malformed json. Error: {str(e)}') from e
 
     return result